requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aioredis==1.3.1
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
autopep8==2.3.2
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
brotlipy==0.7.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
confection==0.1.5
constantly==23.10.4
contourpy==1.3.1
core4ai==1.0.0
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
daphne==3.0.2
databricks-sdk==0.49.0
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
Django==5.2
django-admin-filters==0.1.3
django-admin-interface==0.30.0
django-admin-list-filter-dropdown==1.0.3
django-allauth==65.7.0
django-axes==7.0.2
django-channels==0.7.0
django-colorfield==0.13.0
django-csp==4.0
django-environ==0.12.0
django-ipware==7.0.1
django-jazzmin==3.0.1
django-otp==1.6.0
django-otp-twilio==1.0.5
django-parler==2.3
django-prometheus==2.3.1
django-ratelimit==4.1.0
django-redis==5.4.0
django-rosetta==0.10.2
django-structlog==9.1.1
django-widget-tweaks==1.5.0
docker==7.1.0
docopt==0.6.2
dotenv==0.9.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
fasttext-wheel==0.9.2
filelock==3.18.0
flashtext==2.7
Flask==3.1.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
fuzzywuzzy==0.18.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.71.0
gunicorn==23.0.0
h11==0.14.0
hiredis==3.1.0
httpcore==1.0.8
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
iniconfig==2.1.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
langchain==0.3.23
langchain-core==0.3.51
langchain-mcp-adapters==0.0.7
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langcodes==3.5.0
langgraph==0.3.29
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.30
language_data==1.3.0
Levenshtein==0.27.1
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mcp==1.6.0
mdurl==0.1.2
mistune==3.1.3
mlflow==2.21.3
mlflow-skinny==2.21.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
murmurhash==1.0.12
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.3
nltk==3.9.1
numpy==2.2.4
oauthlib==3.2.2
openai==1.73.0
opentelemetry-api==1.32.0
opentelemetry-sdk==1.32.0
opentelemetry-semantic-conventions==0.53b0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parsivar==0.2.3.1
parso==0.8.4
pexpect==4.9.0
phonenumbers==9.0.4
pickleshare==0.7.5
pillow==11.2.1
pipdeptree==2.26.0
pipreqs==0.5.0
platformdirs==4.3.7
pluggy==1.5.0
polib==1.2.0
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-django==4.11.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-ipware==3.0.0
python-Levenshtein==0.27.1
python-slugify==8.0.4
python-telegram-bot==22.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qrcode==8.1
RapidFuzz==3.13.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
soupsieve==2.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
sqlparse==0.5.3
srsly==2.5.1
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.2
structlog==25.2.0
sympy==1.13.1
tenacity==9.1.2
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
twilio==9.5.2
Twisted==24.11.0
txaio==23.1.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.1
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.5
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.9.0
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarg==0.1.9
yarl==1.19.0
zipp==3.21.0
zope.interface==7.2
zstandard==0.23.0
